+++ docs/changes.rst
@@ -14,6 +14,7 @@ Not yet released.
 
 * Avoid query parser crash in multi-threaded environments.
 * Avoid :ref:`autofix` crash on multi-value strings.
+* Make project tokens work when :ref:`2fa` or :ref:`component-agreement` are enforced.
 
 **Compatibility**